// src/types/infringe/Infringe.ts
type Infringe = {
  id: string;
  content: string;
  apply_for: string;
  apply_for_detail?: string;
  punishment: string;
  detail?: string;
  updated_punishment_detail?: string;
  note?: string;
  note_detail?: string;
  added_punishment?: string;
  added_punishment_detail?: string;
  updated_added_punishment_detail?: string;
  solution?: string;
  solution_detail?: string;
  updated_solution_detail?: string;
};

export default Infringe;